moved token commands to subfolder / added DeleteTokenCommand
