+++ packages/hedera-wallet-snap/packages/snap/src/commands/hts/CreateTokenCommand.ts
@@ -18,7 +18,7 @@
  *
  */
 
-import { TokenCustomFee } from '../types/params';
+import { TokenCustomFee } from '../../types/params';
 import {
   AccountId,
   Client,
@@ -30,9 +30,9 @@ import {
   TokenSupplyType,
   TokenType,
 } from '@hashgraph/sdk';
-import { TxReceipt } from '../types/hedera';
-import { Utils } from '../utils/Utils';
-import { CryptoUtils } from '../utils/CryptoUtils';
+import { TxReceipt } from '../../types/hedera';
+import { Utils } from '../../utils/Utils';
+import { CryptoUtils } from '../../utils/CryptoUtils';
 
 export class CreateTokenCommand {
   readonly #assetType: 'TOKEN' | 'NFT';
