+++ packages/hedera-wallet-snap/packages/snap/src/commands/hts/DissociateTokensCommand.ts
@@ -1,29 +1,9 @@
-/*-
- *
- * Hedera Wallet Snap
- *
- * Copyright (C) 2024 Hedera Hashgraph, LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
+import { AccountId, Client, TokenDissociateTransaction } from '@hashgraph/sdk';
+import { TxReceipt } from '../../types/hedera';
+import { Utils } from '../../utils/Utils';
+import { CryptoUtils } from '../../utils/CryptoUtils';
 
-import { AccountId, Client, TokenAssociateTransaction } from '@hashgraph/sdk';
-import { TxReceipt } from '../types/hedera';
-import { Utils } from '../utils/Utils';
-import { CryptoUtils } from '../utils/CryptoUtils';
-
-export class AssociateTokensCommand {
+export class DissociateTokensCommand {
   readonly #tokenIds: string[];
 
   constructor(tokenIds: string[]) {
@@ -31,7 +11,7 @@ export class AssociateTokensCommand {
   }
 
   public async execute(client: Client): Promise<TxReceipt> {
-    const transaction = new TokenAssociateTransaction()
+    const transaction = new TokenDissociateTransaction()
       .setAccountId(client.operatorAccountId as AccountId)
       .setTokenIds(this.#tokenIds)
       .freezeWith(client);
