+++ packages/hedera-wallet-snap/packages/snap/src/facades/MintTokenFacade.ts
@@ -22,7 +22,7 @@ import { providerErrors } from '@metamask/rpc-errors';
 import { divider, heading, text } from '@metamask/snaps-ui';
 import _ from 'lodash';
 import { HederaClientImplFactory } from '../client/HederaClientImplFactory';
-import { MintTokenCommand } from '../commands/MintTokenCommand';
+import { MintTokenCommand } from '../commands/hts/MintTokenCommand';
 import { TxReceipt } from '../types/hedera';
 import { MintTokenRequestParams } from '../types/params';
 import { SnapDialogParams, WalletSnapParams } from '../types/state';
